requirements.txt
youtube_transcript_api
phidata
google-generativeai
pgvector
sqlalchemy 
psycopg
uvicorn
fastapi
